requirements.txt
aiohttp==3.8.3
python-dotenv==0.21.0
python-magic==0.4.27
python-lsp-server==1.8.0
